% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageRisk.R
\name{ageRisk}
\alias{ageRisk}
\title{Returns the conditional probability mass function of being infected given an age and gender}
\usage{
ageRisk(cases, populationPyramid, gender = TRUE, plot = FALSE)
}
\arguments{
\item{cases}{A data frame with a numeric vector with the ages of cases in years and a vector with gender}

\item{populationPyramid}{A dataframe with the count of individuals}

\item{gender}{= TRUE (default) A boolean to indicate the disaggregation of age risk by gender}

\item{plot}{= FALSE (default) A boolean for displaying a plot}
}
\value{
A dataframe with the proportion or total count of individuals
}
\description{
Function that returns the probability of being infected given an age
}
