% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endemic_channel.R
\name{auto_endemic_channel}
\alias{auto_endemic_channel}
\title{Returns an automated endemic channel of a disease}
\usage{
auto_endemic_channel(
  disease_name,
  divipola_code,
  year,
  observations = NULL,
  location = "O",
  window = 7,
  method = "geometric",
  geom_method = "shifted",
  outlier_years = NULL,
  outliers_handling = "ignored",
  ci = 0.95,
  plot = TRUE
)
}
\arguments{
\item{disease_name}{Disease to be consulted}

\item{divipola_code}{DIVIPOLA code of the municipality where the disease is
consulted}

\item{year}{Year of observations}

\item{observations}{A numeric vector with the current observations (Optional)}

\item{location}{TBD}

\item{window}{A numeric value to specify the number of previous and
posterior periods to include in the calculation of the current period mean}

\item{method}{A string with the mean calculation method of preference
(median, mean, or geometric) or to use unusual behavior method (Poisson
Distribution Test for hypoendemic settings)}

\item{geom_method}{A string with the selected method for geometric mean
calculation, see: geom_mean}

\item{outlier_years}{A numeric vector with the outlier years}

\item{outliers_handling}{A string with the handling decision regarding
outlier years}

\item{ci}{= 0.95 A numeric value to specify the confidence interval to use
with the geometric method}

\item{plot}{A boolean for displaying a plot}
}
\value{
TBD
}
\description{
Function that performs automated data wrangling necessary for
an automated endemic channel given a method and an specific disease, location
and year.
}
\examples{
\dontrun{
unusual_behaviour(incidence_historic)
}

}
