% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_ibague_2014-data.R
\docType{data}
\name{dengue_ibague_2014}
\alias{dengue_ibague_2014}
\title{Epidemiological data of Ibague city in Colombia (2014)}
\format{
linelist
}
\source{
Sivigila, \url{http://portalsivigila.ins.gov.co/buscador}
}
\usage{
data(dengue_ibague_2014)
}
\description{
Linelist dataframe of dengue cases occurred in Ibague city (Colombia)
in 2014
}
\examples{
data(dengue_ibague_2014)
ages_in_years <- dengue_ibague_2014$EDAD_REAL


}
\references{
Prueba
}
\keyword{datasets}
