% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_orinoquia_2017-data.R
\docType{data}
\name{dengue_orinoquia_2017}
\alias{dengue_orinoquia_2017}
\title{Epidemiological data of Orinoquia region in Colombia (2017)}
\format{
linelist
}
\source{
Sivigila, \url{http://portalsivigila.ins.gov.co/buscador}
}
\usage{
data(dengue_orinoquia_2017)
}
\description{
Linelist dataframe of dengue cases occurred in Orinoquia region (Colombia)
in 2017
}
\examples{
data(dengue_orinoquia_2017)
ages_in_years <- dengue_orinoquia_2017$EDAD_REAL

}
\references{
Prueba
}
\keyword{datasets}
