% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_tolima_2014-data.R
\docType{data}
\name{dengue_tolima_2014}
\alias{dengue_tolima_2014}
\title{Epidemiological data of Tolima department in Colombia (2017)}
\format{
linelist
}
\source{
Sivigila, \url{http://portalsivigila.ins.gov.co/buscador}
}
\usage{
data(dengue_tolima_2014)
}
\description{
Linelist dataframe of dengue cases occurred in Tolima department (Colombia)
in 2014
}
\examples{
data(dengue_tolima_2014)
ages_in_years <- dengue_tolima_2014$EDAD_REAL

}
\references{
Prueba
}
\keyword{datasets}
