% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{describe_ethnicity}
\alias{describe_ethnicity}
\title{Provides the sociological description of ethnicities in Colombia}
\usage{
describe_ethnicity(ethniclabels, language = "ES", plot = FALSE)
}
\arguments{
\item{ethniclabels}{A numeric vector with the codes of ethnicities to consult}

\item{language}{"ES" for description in spanish "EN" for english}

\item{plot}{A boolean for displaying an histogram plot}
}
\value{
A printed message with the description of the ethnicities
}
\description{
Function that returns the description of consulted ethnicities
}
\examples{
\donttest{
describe_ethnicity(c(1, 2, 3, 4))
}
}
