% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{describe_occupation}
\alias{describe_occupation}
\title{Get ISCO-88 occupation labels from codes}
\usage{
describe_occupation(isco_codes, gender = NULL, plot = NULL)
}
\arguments{
\item{isco_codes}{A numeric vector of ISCO-88 occupation codes
(major, submajor, minor or unit)}

\item{gender}{A vector with the isco_codes vector genders}

\item{plot}{A type of plot between boxes and circular}
}
\value{
A string vector of ISCO-88 labels
}
\description{
Function that translates a vector of ISCO-88 occupation codes
into a vector of labels
}
\examples{
\dontrun{
describe_occupation(1111, level = 1)
}
}
