% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\docType{data}
\name{distance_matrix}
\alias{distance_matrix}
\title{distance_matrix}
\format{
An object of class \code{data.frame} with 1121 rows and 1121 columns.
}
\usage{
data(distance_matrix)
}
\description{
Distance matrix between all Colombia's municipalities by land
or river.
}
\details{
Distance matrix between municipalities
}
\keyword{datasets}
