% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola_table.R
\docType{data}
\name{divipola_table}
\alias{divipola_table}
\title{divipola_table}
\format{
An object of class \code{data.frame} with 1121 rows and 7 columns.
}
\usage{
divipola_table
}
\description{
Political and administrative distribution of Colombia's
municipalities
}
\keyword{datasets}
