% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endemic_channel.R
\name{endemic_plot}
\alias{endemic_plot}
\title{Creates the endemic channel plot}
\usage{
endemic_plot(channel_data, method, outlier_years, outliers_handling, language)
}
\arguments{
\item{channel_data}{Data frame with the central tendency, upper limit,
lower limit, and observations to plot}

\item{method}{A string with the method used in the endemic channel
calculation}

\item{outlier_years}{A numeric vector with the outlier years}

\item{outliers_handling}{A string with the handling decision regarding
outlier years}

\item{language}{Language for plot components}
}
\value{
The ggplot object with the endemic channel plot
}
\description{
Function that creates the endemic channel plot
}
\keyword{internal}
