% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endemicChannel.R
\name{endemicChannel}
\alias{endemicChannel}
\title{Returns the endemic channel of a disease}
\usage{
endemicChannel(
  observations,
  incidence_historic,
  method = "geometric",
  geom_method = "shifted",
  window = 0,
  outlier_years = NULL,
  outliers_handling = "ignored",
  CI = 0.95,
  plot = FALSE
)
}
\arguments{
\item{observations}{A numeric vector with the current observations}

\item{incidence_historic}{An incidence object with the historic observations}

\item{method}{= "geometric" (default) A string with the mean calculation method of preference (median, mean, or geometric)}

\item{geom_method}{= NULL (default) A string with the selected method for geometric mean calculation, see: geomMean}

\item{window}{= 0 (default) A numeric value to specify the number of previous and posterior periods to include in the calculation of the current period mean}

\item{outlier_years}{= NULL (default) A numeric vector with the oulier years}

\item{outliers_handling}{= "ignore" (default) A string with the handling decision regarding outlier years
\itemize{
\item ignored = data from outlier years will not take into account for the calculations
\item included = data from outlier years will take into account for the calculations
\item replaced_by_median = data from outlier years will be replaced with the median and take into account for the calculations
\item replaced_by_mean = data from outlier years will be replaced with the mean and take into account for the calculations
\item replaced_by_geom_mean = data from outlier years will be replaced with the geometric mean and take into account for the calculations
}}

\item{CI}{= 0.95 (default) A numeric value to specify the confidence interval to use with the geometric method}

\item{plot}{= FALSE (default) A boolean for displaying a plot}
}
\value{
A dataframe with the observation, historical mean, and confidence intervals (or risk areas)
}
\description{
Function that builds the endemic channel of a disease time series based on the selected method and windows of observation
}
\examples{
endemicChannel(observations, incidence_historic, method = "geometric", outliers_handling = "replace_with_median", plot = TRUE)
}
