% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{epi_calendar}
\alias{epi_calendar}
\title{Get the epidemiological calendar of a consulted year.}
\usage{
epi_calendar(year, jan_days = 4)
}
\arguments{
\item{year}{A numeric value for the year of interest.}

\item{jan_days}{Number of January days that the first epidemiological week
must contains.}
}
\value{
A character array with the starting dates of the epidemiological
weeks of the given year.
}
\description{
Function that returns the starting date of the epidemiological
weeks in a year of interest.
}
\examples{
epi_calendar(2016)

}
