% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{epi_georef}
\alias{epi_georef}
\title{Complete georeference data from an epidemiologic linelist.}
\usage{
epi_georef(query_vector)
}
\arguments{
\item{query_vector}{A vector of DIVIPOLA codes, names or coordinates.}
}
\value{
A dataframe with DIVIPOLA codes, names, longitude and latitude from
the query.
}
\description{
function that complete a structure of municipality names,
DIVIPOLA codes, and coordinates from an input of one of these variables.
}
\examples{
\dontrun{
epigeoref(query_vector)
}

}
