% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiCalendar.R
\name{epiCalendar}
\alias{epiCalendar}
\title{Get the epidemiological calendar of a consulted year}
\usage{
epiCalendar(year, daysJan = 4)
}
\arguments{
\item{year}{A numeric value for the year of interest}

\item{daysJan}{= 4 (default) Number of January days that the first epidemiological week must contains}
}
\value{
A character array with the starting dates of the epidemiological weeks
}
\description{
Function that returns the starting date of the epidemiological weeks in a year of interest
}
\examples{
epiCalendar(2016)
}
