% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_projection_col.R,
%   R/spatial_polygons_col.R
\docType{data}
\name{population_projection_col_0}
\alias{population_projection_col_0}
\alias{spatial_polygons_col_0}
\title{population_projection_col}
\format{
An object of class \code{data.frame} with 22 rows and 310 columns.

An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 2 columns.
}
\usage{
population_projection_col_0

spatial_polygons_col_0
}
\description{
Documentation for datasets regarding population projection

Population projection of Colombia.

Documentation for datasets regarding spatial polygons

Spatial polygon of Colombia.
}
\keyword{datasets}
