% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_projection_col.R
\name{epiCo-population-data}
\alias{epiCo-population-data}
\title{population_projection_col}
\description{
Documentation for datasets regarding population projection
}
\details{
Colombian Population projection
}
