% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigeoref.R
\name{epigeoref}
\alias{epigeoref}
\title{Complete georeference data from an epidemiologic linelist}
\usage{
epigeoref(query_vector)
}
\arguments{
\item{query}{A vector of DIVIPOLA codes, names or coordinates}
}
\value{
A dataframe with DIVIPOLA codes, names, longitude and latitude from the query
}
\description{
Function that complete a structure of municipality names, DIVIPOLA codes, and coordinates
from an input of one of these variables
}
\examples{
epigeoref(query_vector)
}
