% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_incidenceRate.R
\name{estimate_incidenceRate}
\alias{estimate_incidenceRate}
\title{Extends an incidence class object with incidence rates estimations}
\usage{
estimate_incidenceRate(incidence_object, level, scale = 1e+05)
}
\arguments{
\item{level}{Administration level at which incidence counts are grouped (0=national, 1=state/department, 2=city/municipality)}

\item{incidence}{An incidence object}

\item{country}{3 letter ISO code for the country of interest}
}
\value{
A modified incidence object where counts are normalized with population
}
\description{
Function that estimates incidence rates from a incidence class object and population projections
}
