% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicity.R
\name{ethnicity}
\alias{ethnicity}
\title{Provides the sociological description of ethnicities in Colombia}
\usage{
ethnicity(ethniclabels, language = "ES", plot = FALSE)
}
\arguments{
\item{ethniclabels}{A numeric vector with the codes of ethnicities to consult}

\item{plot}{= FALSE (defalut) A boolean for displaying a plot}
}
\value{
A printed message with the description of the ethnicities
}
\description{
Function that returns the description of consulted ethnicities
}
\examples{
ethnicity(c(1,2,3,4))
}
