% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{get_age_risk_sex}
\alias{get_age_risk_sex}
\title{Auxiliary function to calculate the proportion by age according to the total
population and sex}
\usage{
get_age_risk_sex(age, sex_vector, pyramid, sex)
}
\arguments{
\item{age}{A vector with the ages of cases in years from 0 to 100 years}

\item{sex_vector}{A vector with the sex of cases 'F' and 'M'}

\item{pyramid}{A dataframe with the count of individuals}

\item{sex}{A string specifying the sex being calculated}
}
\value{
A dataframe with the proportion by age according to the total
population and sex
}
\description{
Auxiliary function to calculate the proportion by age according to the total
population and sex
}
\keyword{internal}
