% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Estimates the mode of an input vector}
\usage{
get_mode(x)
}
\arguments{
\item{x}{A numeric or character vector}
}
\value{
The mode of the vector
}
\description{
Function that estimates the mode of an input vector
}
\examples{
get_mode(x)
}
