% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occupationLabels.R
\name{get_occupationLabels}
\alias{get_occupationLabels}
\title{Get ISCO-88 occupation labels from codes}
\usage{
get_occupationLabels(isco_codes, output_level)
}
\arguments{
\item{isco_codes}{A numeric vector of ISCO-88 occupation codes (major, submajor, minor or unit)}

\item{output_level}{A string parameter that defines the level of the desired label (major, submajor, minor or unit)}
}
\value{
A string vector of ISCO-88 labels
}
\description{
Function that translates a vector of ISCO-88 occupation codes into a vector of labels
}
\examples{
estimate_incidenceRate(incidence, level = 1)
}
