% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco88_table.R
\docType{data}
\name{isco88_table}
\alias{isco88_table}
\title{isco88_table}
\format{
An object of class \code{data.frame} with 390 rows and 8 columns.
}
\usage{
data(isco88_table)
}
\description{
ISCO88 description of occupations
}
\details{
ISCO88 occupation table
}
\keyword{datasets}
