% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{neighborhoods}
\alias{neighborhoods}
\title{Get ISCO-88 occupation labels from codes}
\usage{
neighborhoods(codes = c("all"), threshold = 2)
}
\arguments{
\item{codes}{Codes of the municipalities to consider for the neighborhoods.}

\item{threshold}{Maximum traveling time around each city}
}
\value{
neighborhood object according to the introduced threshold
}
\description{
Function that translates a vector of ISCO-88 occupation codes into a vector
of labels.
}
\examples{
neighborhood(codes, 2)
}
