% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{occupation_plot}
\alias{occupation_plot}
\title{Distribution plots for ISCO-88 occupation labels}
\usage{
occupation_plot(isco_codes, gender = NULL)
}
\arguments{
\item{isco_codes}{A numeric vector of ISCO-88 occupation codes (unit level)}

\item{gender}{A vector with the gender of cases 'F' and 'M'}
}
\value{
A plot to summarize the distribution of ISCO-88 labels
}
\description{
Function that plot a vector of ISCO-88 occupation codes
}
\examples{
\dontrun{
occupation_plot(1111, level = 1)
}
}
