% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{occupation_plot_circular}
\alias{occupation_plot_circular}
\title{Distribution plots for ISCO-88 occupation labels}
\usage{
occupation_plot_circular(occupation_data, q = 0.9)
}
\arguments{
\item{occupation_data}{A dataframe}

\item{q}{A number that represents the quantile. The default value is 0.9}
}
\value{
A plot to summarize the distribution of ISCO-88 labels
}
\description{
Function that makes a circular packing plot of a vector of
ISCO-88 occupation codes
}
\keyword{internal}
