% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_projection_col.R
\docType{data}
\name{population_projection_col_1}
\alias{population_projection_col_1}
\alias{population_projection_col_2}
\title{Department population projection}
\format{
An object of class \code{data.frame} with 726 rows and 308 columns.

An object of class \code{data.frame} with 24684 rows and 308 columns.
}
\usage{
data(population_projection_col_1)

data(population_projection_col_2)
}
\description{
Population projection of the departments of Colombia.

Population projection of the municipalities of Colombia.
}
\keyword{datasets}
