% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_projection_col.R
\docType{data}
\name{population_projection_col_0}
\alias{population_projection_col_0}
\title{Colombian population projection}
\format{
An object of class \code{data.frame} with 22 rows and 310 columns.
}
\usage{
data(population_projection_col_0)
}
\description{
Population projection of Colombia.
}
\keyword{datasets}
