% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{population_pyramid}
\alias{population_pyramid}
\title{Returns the population pyramid of the consulted region}
\usage{
population_pyramid(
  divipola_code,
  year,
  gender = TRUE,
  range = 5,
  total = TRUE,
  plot = FALSE
)
}
\arguments{
\item{divipola_code}{A numeric code accounting for the territory of interest}

\item{year}{A numeric input for the year of interest}

\item{gender}{A boolean to consult data disaggregated by gender}

\item{range}{A numeric value from 1 to 100 for the age range to use}

\item{total}{A boolean for returning the total number rather than the
proportion of the country's population}

\item{plot}{A boolean for displaying a plot}
}
\value{
A dataframe with the proportion or total count of individuals
}
\description{
Function that returns the population pyramid of the municipality
or department of a specific year
}
\examples{
population_pyramid(15001L, 2015, gender = TRUE, total = TRUE, plot = TRUE)
}
