% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{population_pyramid_plot}
\alias{population_pyramid_plot}
\title{Returns the population pyramid plot}
\usage{
population_pyramid_plot(pop_pyramid, gender = TRUE, total = TRUE)
}
\arguments{
\item{pop_pyramid}{A dataframe with the age counts}

\item{gender}{A boolean to consult data disaggregated by gender}

\item{total}{A boolean for returning the total number rather than the
proportion of the country's population}
}
\value{
the population pyramid plot
}
\description{
Function that returns the population pyramid plot of the
municipality or department of a specific year
}
\keyword{internal}
