% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationPyramid.R
\name{populationPyramid}
\alias{populationPyramid}
\title{Returns the population pyramid of the consulted region}
\usage{
populationPyramid(
  DIVIPOLA_code,
  year,
  gender = TRUE,
  total = TRUE,
  plot = FALSE
)
}
\arguments{
\item{DIVIPOLA_code}{A numeric code accounting for the territory of interest}

\item{year}{A numeric input for year of interest}

\item{gender}{= TRUE (default) A boolean to indicate that data is disaggregated by gender}

\item{total}{= TRUE (default) A boolean for returning the total number rather than the porportion of the populations}

\item{plot}{= FALSE (default) A boolean for displaying a plot}
}
\value{
A dataframe with the proportion or total count of individuals
}
\description{
Function that returns the population pyramid of the municipality or department of an specific year
}
\examples{
populationPyramid(15001, 2015, total = TRUE, plot = TRUE)
}
