% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_polygons_col.R
\docType{data}
\name{spatial_polygons_col_0}
\alias{spatial_polygons_col_0}
\alias{spatial_polygons_col_1}
\alias{spatial_polygons_col_2}
\title{Colombian spatial polygons}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 2 columns.

An object of class \code{SpatialPolygonsDataFrame} with 33 rows and 9 columns.

An object of class \code{SpatialPolygonsDataFrame} with 1121 rows and 12 columns.
}
\usage{
data(spatial_polygons_col_0)

data(spatial_polygons_col_1)

data(spatial_polygons_col_2)
}
\description{
Spatial polygon of Colombia.

Spatial polygons of the departments of Colombia.

Spatial polygons of the municipalities of Colombia.
}
\keyword{datasets}
