% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{age_risk}
\alias{age_risk}
\title{Returns the probability mass function of being infected given age and gender}
\usage{
age_risk(age, gender = NULL, population_pyramid, plot = FALSE)
}
\arguments{
\item{age}{A vector with the ages of cases in years from 0 to 100 years}

\item{gender}{A vector with the gender of cases 'F' and 'M'}

\item{population_pyramid}{A dataframe with the count of individuals}

\item{plot}{A boolean for displaying a plot}
}
\value{
A dataframe with the proportion or total count of individuals
}
\description{
Function that returns the probability of being infected given
age and gender
}
