% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{describe_ethnicity}
\alias{describe_ethnicity}
\title{Provides the sociological description of ethnicities in Colombia}
\usage{
describe_ethnicity(ethnic_labels, language = "ES")
}
\arguments{
\item{ethnic_labels}{A numeric vector with the codes of ethnicities to
consult}

\item{language}{"ES" for description in spanish "EN" for english}
}
\value{
A printed message with the description of the ethnicities
}
\description{
Function that returns the description of consulted ethnicities
}
\examples{
\dontrun{
describe_ethnicity(c(1, 2, 3, 4))
}
}
