% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{describe_ethnicity}
\alias{describe_ethnicity}
\title{Provides the sociological description of ethnicities in Colombia}
\usage{
describe_ethnicity(ethnic_codes, language = c("EN", "ES"))
}
\arguments{
\item{ethnic_codes}{A numeric vector with the codes of ethnicities to
consult}

\item{language}{"ES" for description in Spanish "EN" for English. The default
value is EN}
}
\value{
A printed message with ethnicities descriptions
}
\description{
Function that returns the description of the consulted
ethnicities
}
\examples{
describe_ethnicity(round(runif(n = 150, min = 1, max = 4)))
}
