% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma.R
\name{detect_outbreaks_ewma}
\alias{detect_outbreaks_ewma}
\title{Detects Outbreaks with SPC schemes}
\usage{
detect_outbreaks_ewma(epidata, lambda, nsigmas)
}
\arguments{
\item{epidata}{A data frame with the incidence or incidence rate time series}

\item{lambda}{An smoothing parameter 0 <= lambda <= 1}

\item{nsigmas}{The numeric value of sigmas to use for control limits}
}
\value{
A list with Y smoothed values and limits UCL & LCL
}
\description{
Function that performs EWMA Statistical Process Control to detect outliers
}
