% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_data.R
\docType{data}
\name{epi_data}
\alias{epi_data}
\title{epi_data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66747 rows and 16 columns.
}
\usage{
data(epi_data)
}
\description{
Epidemiological data for the Tolima department for the years
2012 to 2022
}
\details{
Epidemiological data
}
\keyword{datasets}
