% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiCo-package.R
\docType{package}
\name{epiCo-package}
\alias{epiCo}
\alias{epiCo-package}
\title{epiCo: Statistical and Viz Tools for Vector-Borne Diseases in Colombia}
\description{
Provides statistical and visualization tools for the analysis of demographic indicators, and spatio-temporal behavior and characterization of outbreaks of vector-borne diseases (VBDs) in Colombia. It implements travel times estimated in Bravo-Vega C., Santos-Vega M., & Cordovez J.M. (2022), and the endemic channel method (Bortman, M. (1999) \url{https://iris.paho.org/handle/10665.2/8562}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epiverse-trace.github.io/epiCo/}
  \item \url{https://github.com/epiverse-trace/epiCo}
  \item Report bugs at \url{https://github.com/epiverse-trace/epiCo/issues}
}

}
\author{
\strong{Maintainer}: Juan D. Umaña \email{jd.umana10@uniandes.edu.co} (\href{https://orcid.org/0000-0003-0316-6164}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Juan Montenegro-Torres \email{jf.montenegro@uniandes.edu.co} (\href{https://orcid.org/0000-0003-2755-4743}{ORCID})
  \item Julian Otero \email{jd.otero10@uniandes.edu.co} (\href{https://orcid.org/0009-0006-0429-7747}{ORCID})
}

Other contributors:
\itemize{
  \item Hugo Gruson \email{hugo.gruson+R@normasup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor]
}

}
\keyword{internal}
