% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{incidence_rate}
\alias{incidence_rate}
\title{Extends an incidence class object with incidence rates estimations.}
\usage{
incidence_rate(incidence_object, level, scale = 1e+05)
}
\arguments{
\item{level}{Administration level at which incidence counts are grouped.}

\item{scale}{Scale to consider when calculating the incidence_rate.
(0=national, 1=state/department, 2=city/municipality).}

\item{incidence}{An incidence object.}
}
\value{
A modified incidence object where counts are normalized with the
population.
}
\description{
Function that estimates incidence rates from a incidence class
object and population projections.
}
\examples{
\dontrun{
  incidence_rate(incidence_object, 2)
}

}
