% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemporal.R
\name{morans_index}
\alias{morans_index}
\title{Calculate spatial correlation of given municipalities in an incidence_rate
object.}
\usage{
morans_index(
  incidence_object,
  level,
  scale = 1e+05,
  threshold = 2,
  plot = TRUE
)
}
\arguments{
\item{threshold}{Maximum traveling time around each municipality.}

\item{plot}{if TRUE, returns a plot of influential observations in the
Moran's plot.}

\item{incidence_rate}{Incidence rate object with only one observation for a
group of municipalities.}
}
\value{
List of Moran's I clustering analysis, giving the quadrant of each
observation, influential values.
}
\description{
Function to calculate spatial autocorrelation via Moran's Index
from a given incidence_rate object grouped by municipality.
}
\examples{
\dontrun{
morans_index(incidence_rate, 2, FALSE)
}
}
