% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{occupation_plot}
\alias{occupation_plot}
\title{Distribution plots for ISCO-88 occupation labels}
\usage{
occupation_plot(occupation_data, sex = FALSE, q = 0.9)
}
\arguments{
\item{occupation_data}{A dataframe}

\item{sex}{A boolean for sex data. The default value is FALSE}

\item{q}{A number that represents the quantile. The default value is 0.9}
}
\value{
A plot to summarize the distribution of ISCO-88 labels
}
\description{
Function that makes a treemap plot of a vector of ISCO-88
occupation codes
}
\keyword{internal}
