% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{occupation_plot_2}
\alias{occupation_plot_2}
\title{Distribution plots for ISCO-88 occupation labels}
\usage{
occupation_plot_2(occupation_data, gender = FALSE, q = 0.9)
}
\arguments{
\item{occupation_data}{A dataframe}

\item{gender}{A boolean for gender data}
}
\value{
A plot to summarize the distribution of ISCO-88 labels
}
\description{
Function that plot a vector of ISCO-88 occupation codes
}
\examples{
\dontrun{
occupation_plot(1111, level = 1)
}
}
