% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{population_pyramid_plot}
\alias{population_pyramid_plot}
\title{Returns the population pyramid plot}
\usage{
population_pyramid_plot(pop_pyramid, language, sex = TRUE)
}
\arguments{
\item{pop_pyramid}{A dataframe with the age counts}

\item{language}{Language for plot components}

\item{sex}{A boolean to consult data disaggregated by sex. The default value
is TRUE}
}
\value{
the population pyramid plot
}
\description{
Function that returns the population pyramid plot of the
municipality or department of a specific year
}
\keyword{internal}
